Add some validity testing for AccessBoundaryRules and add documentation.
