+++ google/downscope/downscoping_test.go
@@ -1,4 +1,4 @@
-package downscoped
+package downscope
 
 import (
 	"context"
